Moving files from ParameterEditors and ValueConverters folders under PropertyEditors
