+++ src/Umbraco.Infrastructure/PropertyEditors/ParameterEditors/PropertyGroupParameterEditor.cs
@@ -1,5 +1,4 @@
-﻿using Umbraco.Web.Composing;
-using Umbraco.Core.Logging;
+﻿using Umbraco.Core.Logging;
 using Umbraco.Core.PropertyEditors;
 using Umbraco.Core.Services;
 using Umbraco.Core.Strings;
@@ -7,13 +6,13 @@ using Umbraco.Core.Strings;
 namespace Umbraco.Web.PropertyEditors.ParameterEditors
 {
     [DataEditor(
-        "contentTypeMultiple",
+        "tabPicker",
         EditorType.MacroParameter,
-        "Multiple Content Type Picker",
+        "Tab Picker",
         "entitypicker")]
-    public class MultipleContentTypeParameterEditor : DataEditor
+    public class PropertyGroupParameterEditor : DataEditor
     {
-        public MultipleContentTypeParameterEditor(
+        public PropertyGroupParameterEditor(
             ILogger logger,
             IDataTypeService dataTypeService,
             ILocalizationService localizationService,
@@ -22,8 +21,10 @@ namespace Umbraco.Web.PropertyEditors.ParameterEditors
             : base(logger, dataTypeService, localizationService, localizedTextService, shortStringHelper)
         {
             // configure
-            DefaultConfiguration.Add("multiple", true);
-            DefaultConfiguration.Add("entityType", "DocumentType");
+            DefaultConfiguration.Add("multiple", "0");
+            DefaultConfiguration.Add("entityType", "PropertyGroup");
+            //don't publish the id for a property group, publish it's alias (which is actually just it's lower cased name)
+            DefaultConfiguration.Add("publishBy", "alias");
         }
     }
 }
